pyproject.toml
[project]
name = "inference-gateway-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.133.0",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "pydantic-settings>=2.13.1",
    "uvicorn[standard]>=0.35.0",
]
